Kernel: Add initial polling support for NVMe
Still the admin queue uses IRQ to send completion but we ignore by not
enabling the irq if in polled mode.
NVMeQueue's base class is IRQhandler which automatically calls irq
specific function which might cause failures in hardware. Call disable
irq only when irq is registered might avoid that issue.