Kernel/NVMeQueue: Use waitqueue to in submit_sync_sqe
The current way we handle sync commands is very ugly and depends on lot
of preconditions. Now that we have a end_io handler for a request, we
can use WaitQueue to do sync commands more elegantly.

This does depend on block layer sending one request at a time but this
change is a step forward towards better IO handling.